Alter FeatureVector constructor to take a set instead of a list.
Some asserts, such as in updateLength(), fail if a vector contains
duplicate indices. This happens when, e.g., a backoff feature turns out
to the be same as a finer feature. The simplest way to avoid this is to
have the feature functions pass a set (which ensures uniqueness) instead
of a list.