Enable training with several different hyperparameter values.
The purpose of this is to avoid recomputing the fst for each example
prior to training. Now, we compute the fsts once, and since the features
do not change, we can reuse the fsts for training weight vectors using
multiple regularization constants and tolerance thresholds.

The new --beta and --tolerance options can occur multiple times on
the command line. This means that the subclasses of Optimizer no longer
handle these options.

Also renamed some of the main options (e.g., "dir" to "directory") to
their full names, since boost will match partial strings in any event.