Implement edit ops that are conditioned on character classes.
The user can now specify, via a regular expression, a class of
characters that is to be distinguished from the remaining characters in
the dataset's alphabet. We now replace each character-agnostic edit
operation (e.g, Delete) with several operations (e.g., DeleteClass and
DeleteOther) that specify the type of character(s) consumed.