Fix the collapsed features so they work with digit-based alphabets.
The WordPairReader class, for example, reads sequences of numbers. Each
number is the index of a unicode character in some external dictionary,
but the point is that we use this coding to avoid having to deal with
unicode (at least for now). A problem is that if we concatenate numbers
the way we normally concatenate letters, the numbers lose their meaning
(i.e., we can't recover the letters they represent). This commit fixes
this problem in WordAlignmentFeatureGen; it also removes zero order
collapsed features, which are redundant.