Replaced AlignmentTransducer with AlignmentHypergraph throughout.
Portions of unit tests involving max-margin objectives are failing. Can we
attribute this to breaking ties between optimal alignments? It is just a
coincidence that the test involving printAlignment() passes? I suppose the
{stress, sutoressu} pair has a unique max-scoring alignment? Note that the
objective values returned by optimizers are higher under the new regime.
Maybe inspect the output of the opimizer to see what's going on?