Removed - operators from LogWeight and re-implemented / operator.
The subtraction (-) operators were negating the log domain value, which happens
to produce the correct behaviour when multiplication and division are involved
(e.g., a times (-b) = a - b), but not when subtraction is involved because in
this case we would need to maintain a sign bit. Since there is no necessity at
this time to store negative values in the log domain, I decided to avoid
potential misuse of the - operators by removing them entirely and forcing the
calling code to explicitly cast to a double if a log domain value needs to
be negated.