Fixed a bug in computing the feature covariance matrix.
Previously, I was computing the covariance matrix in the log domain and
then exponentiating the result. However, due to a deficiency in the LogWeight
class, this produced an incorrect result in the subtraction step. Instead, I
now exponentiate the feature coocurrences and expectations, then compute the
covariance matrix in the real domain.