Decreased the memory consumption of the u-w objectives.
Apparently, calling clear() for the vector_of_vector type doesn't free all
the memory (maybe the rows are not cleared?). As a consequence, some of
these matrices were accumulating features for each pass that was made through
the data. This was remedied by re-creating the problematic matrices each time
we process an example.