Fixed a bug in RegularizerSoftTying that was causing options to be ignored.
The setBeta function was overriding beta values that were set via command-line
options. The new behaviour is that if any of the four beta values are set, they
must all be set, or else an error is reported. The setBeta function will only
set default (relative) beta values if no options were processed previously.
Finally, the classification parameters are no longer tied by default; that is,
we set _betaSharedW to zero by default.