Trying an idea to fix the shared parameters problem, but it cannot work.
I explicitly added shared_w and shared_u members to Parameters, instead of
increasing the size of w and u to accomodate the shared features. However, when
we go to compute the gradient, the gradient vector is the size of the original
feature space (without the shared ones); so, we have nowhere to store the
shared features gradient!