Changed gradient vectors from RealVec to SparseRealVec.
The impetus for this change was to make SGD faster. Previously, although the
gradient vectors returned by valueAndGradient() are sparse for a single
example or minibatch, the type was RealVec and so SGD would still have to
iterate over every coordinate when performing a parameter update. Now, with
SparseRealVec, it can iterate over the non-zero coordinates only. However, it
is likely that the batch optimizer will now be slower because they will be
storing a dense gradient vector in a sparse data structure. My plan is to
mitigate this by resizing the sparse vector to the dimensionality of the
parameter space before accumulating the gradients.