Fixed a bug where the eval k-best lists could differ from those of train.
That is, instead of using the same initial weights to populate the k-best lists
for the eval data as were used for the train data, the eval data lists were
based on the trained parameters. This is arguably a valid thing to do, so
perhaps it can be enabled as an option in a later version.